Do not delete previous startpage after first use (#17)
* git ignore self created core/app subdirectory

* added checkboxes on Columns page for users wanting to preserve their previously generated startpage

* removed uncessary <pre> tag, moved $_POST debug near $_POST parsing

* excluding new checkboxes from being considered as table keys, along with singlebutton

* generate_start() now has params to keep startpage and to append links

* prevent previous startpage deletion when keep_startpage is checked, append links to previous page if necessary, avoid duplicates

* "startpage page" sounded kind of repetitive

* make sure that a previous startpage was created before trying to keep it alive

+++ core/generate.php
@@ -1,9 +1,3 @@
-<pre>
-    <?php
-      //print_r($_POST);
-    ?>
-</pre>
-
 <?php
 
 require "app/config.php";
@@ -17,6 +11,9 @@ $index_table_rows = '';
 $index_table_headers = '';
 $start_page = '';
 $sort = '';
+$excluded_keys = array('singlebutton', 'keep_startpage', 'append_links');
+$generate_start_checked_links = array();
+$startpage_filename = "app/index.php";
 
 function column_type($columnname){
     switch ($columnname) {
@@ -52,13 +49,60 @@ function generate_error(){
     echo "Generating Error file<br><br>";
 }
 
-function generate_start($start_page){
+function generate_start($start_page, $keep_startpage, $append_links){
     global $startfile;
-    $step0 = str_replace("{TABLE_BUTTONS}", $start_page, $startfile);
-    $destination_file = fopen("app/index.php", "w") or die("Unable to open file!");
-    fwrite($destination_file, $step0);
-    fclose($destination_file);
-    echo "Generating Startpage file<br>";
+    global $generate_start_checked_links;
+    global $startpage_filename;
+
+    // make sure that a previous startpage was created before trying to keep it alive
+    if (!$keep_startpage || ($keep_startpage && !filesize($startpage_filename))) {
+        $step0 = str_replace("{TABLE_BUTTONS}", $start_page, $startfile);
+        $destination_file = fopen($startpage_filename, "w") or die("Unable to open fresh startpage file!");
+        fwrite($destination_file, $step0);
+        fclose($destination_file);
+        echo "Generating Startpage file<br>";
+    } else {
+        if ($append_links) {
+            // load existing template
+            echo "Retrieving existing Startpage file<br>";
+            $handle = fopen($startpage_filename, "r") or die("Unable to open existing startpage file!");;
+            $startfile = fread($handle, filesize($startpage_filename));
+            fclose($handle);
+            
+            // extract existing links from app/index.php
+            echo "Looking for new links to append to Startpage file<br>";
+            $link_matcher_pattern = '/href=["\']?([^"\'>]+)["\']?/im';
+            preg_match_all($link_matcher_pattern, $startfile, $startfile_links);
+            if (count($startfile_links)) {
+                foreach($startfile_links[1] as $startfile_link) {
+                    // echo '- Found existing link '.$startfile_link.'<br>';
+                }
+            }
+
+            // do not append links to app/index.php if they  already 
+            preg_match_all($link_matcher_pattern, $start_page, $start_page_links);
+            if (count($start_page_links)) {
+                foreach($start_page_links[1] as $start_page_link) {
+                    if (!in_array($start_page_link, $generate_start_checked_links)) {
+                        if (in_array($start_page_link, $startfile_links[1])) {
+                            echo '- Not appending '.$start_page_link.' as it already exists<br>';
+                        } else {
+                            echo '- Appending '.$start_page_link.'<br>';
+                            array_push($startfile_links[1], $start_page_link);
+                            $linkname = str_replace('-index.php', '', basename($start_page_link));
+                            $step0 = preg_replace('/<\/div>.*<\/center>/msx', "\t".'<a href="'.$start_page_link.'" class="btn btn-primary" role="button">'.$linkname.'</a>'."\n</div>\n</center>", $startfile);
+                            $destination_file = fopen($startpage_filename, "w") or die("Unable to open file!");
+                            fwrite($destination_file, $step0);
+                            fclose($destination_file);
+                        }
+                        array_push($generate_start_checked_links, $start_page_link);
+                    }
+                }
+            }
+        }
+    }
+
+    
 }
 
 function generate_index($tablename,$tabledisplay,$index_table_headers,$index_table_rows,$column_id, $columns_available, $index_sql_search) {
@@ -135,9 +179,10 @@ function generate_update($tablename, $create_records, $create_err_records, $crea
 }
 
 function count_index_colums($table) {
+    global $excluded_keys;
     $i = 0;
     foreach ( $_POST as $key => $value) {
-        if ($key == 'singlebutton') {
+        if (in_array($key, $excluded_keys)) {
             //echo "nope";
         }
         else if ($key == $table) {
@@ -154,6 +199,10 @@ function count_index_colums($table) {
     }
     return $i;
 }
+
+// echo "<pre>";
+// print_r($_POST);
+// echo "</pre>";
 // Go trough the POST array
 // Every table is a key
 foreach ($_POST as $key => $value) {
@@ -182,7 +231,7 @@ foreach ($_POST as $key => $value) {
     $update_sql_id = '';
     $update_column_rows = '';
 
-    if ($key != 'singlebutton') {
+    if (!in_array($key, $excluded_keys)) {
         $i = 0;
         $j = 0;
         $max = count_index_colums($key)+1;
@@ -424,7 +473,7 @@ foreach ($_POST as $key => $value) {
                     $start_page .= "\n\t";
                 }
 
-                generate_start($start_page);
+                generate_start($start_page, isset($_POST['keep_startpage']) && $_POST['keep_startpage'] == 'true' ? true : false, isset($_POST['append_links']) && $_POST['append_links'] == 'true' ? true : false);
                 generate_error();
                 generate_index($tablename,$tabledisplay,$index_table_headers,$index_table_rows,$column_id, $columns_available,$index_sql_search);
                 generate_create($tablename,$create_records, $create_err_records, $create_sqlcolumns, $create_numberofparams, $create_sql_params, $create_html, $create_postvars);
